+++ src/engines/niseci/x2/mod.rs
@@ -155,10 +155,10 @@ pub fn calculate_x2(
 
     let mut specie_campionate_set: HashMap<&str, bool> = HashMap::new();
     for cattura in campionamento {
-        if (cattura.specie.specie_attesa() == require_specie_attesa)
-            && (cattura.specie.tipo_autoctono() == 1 || cattura.specie.tipo_autoctono() == 2)
+        if (cattura.specie_attesa() == require_specie_attesa)
+            && (cattura.tipo_autoctono() == 1 || cattura.tipo_autoctono() == 2)
         {
-            match specie_campionate_set.entry(cattura.specie.id()) {
+            match specie_campionate_set.entry(cattura.id()) {
                 Entry::Occupied(_) => {}
                 Entry::Vacant(vacant_entry) => {
                     vacant_entry.insert(true);
@@ -208,8 +208,8 @@ pub fn calculate_x2_per_alloctone(
 
     let mut specie_campionate_set: HashMap<&str, bool> = HashMap::new();
     for cattura in campionamento {
-        if cattura.specie.tipo_alloctono() > 0 {
-            match specie_campionate_set.entry(cattura.specie.id()) {
+        if cattura.tipo_alloctono() > 0 {
+            match specie_campionate_set.entry(cattura.id()) {
                 Entry::Occupied(_) => {}
                 Entry::Vacant(vacant_entry) => {
                     vacant_entry.insert(true);
@@ -281,10 +281,10 @@ fn calculate_sommatoria_x2_a(
 
     // riempo l'hashmap con solo le specie autoctone campionate
     for cattura in c {
-        if (cattura.specie.specie_attesa() == require_specie_attesa)
-            && (cattura.specie.tipo_autoctono() == 1 || cattura.specie.tipo_autoctono() == 2)
+        if (cattura.specie_attesa() == require_specie_attesa)
+            && (cattura.tipo_autoctono() == 1 || cattura.tipo_autoctono() == 2)
         {
-            match classi_eta_map.entry(cattura.specie.id()) {
+            match classi_eta_map.entry(cattura.id()) {
                 Entry::Occupied(mut entry) => {
                     entry.get_mut().update_classi_eta(cattura);
                 }
@@ -312,17 +312,17 @@ fn calculate_sommatoria_x2_b(
         HashMap::with_capacity(10);
 
     for cattura in c {
-        if (cattura.specie.specie_attesa() == require_specie_attesa)
-            && (cattura.specie.tipo_autoctono() == 1 || cattura.specie.tipo_autoctono() == 2)
+        if (cattura.specie_attesa() == require_specie_attesa)
+            && (cattura.tipo_autoctono() == 1 || cattura.tipo_autoctono() == 2)
         {
-            match esemplari_per_cattura_map.entry(cattura.specie.id()) {
+            match esemplari_per_cattura_map.entry(cattura.id()) {
                 Entry::Occupied(mut occupied_entry) => {
                     occupied_entry
                         .get_mut()
                         .fill_passaggio(cattura.passaggio_cattura);
                 }
                 Entry::Vacant(vacant_entry) => {
-                    let mut epc = EsemplariPerCattura::new(&cattura.specie);
+                    let mut epc = EsemplariPerCattura::new(cattura.specie());
                     epc.fill_passaggio(cattura.passaggio_cattura);
                     vacant_entry.insert(epc);
                 }
@@ -343,8 +343,8 @@ fn calculate_sommatoria_x2_a_per_alloctone(
 
     // riempo l'hashmap con solo le specie autoctone campionate
     for cattura in c {
-        if cattura.specie.tipo_alloctono() > 0 {
-            match classi_eta_map.entry(cattura.specie.id()) {
+        if cattura.tipo_alloctono() > 0 {
+            match classi_eta_map.entry(cattura.id()) {
                 Entry::Occupied(mut entry) => {
                     entry.get_mut().update_classi_eta(cattura);
                 }
@@ -381,15 +381,15 @@ fn calculate_sommatoria_x2_b_per_alloctone(
         HashMap::with_capacity(10);
 
     for cattura in c {
-        if cattura.specie.tipo_alloctono() > 0 {
-            match esemplari_per_cattura_map.entry(cattura.specie.id()) {
+        if cattura.tipo_alloctono() > 0 {
+            match esemplari_per_cattura_map.entry(cattura.id()) {
                 Entry::Occupied(mut occupied_entry) => {
                     occupied_entry
                         .get_mut()
                         .fill_passaggio(cattura.passaggio_cattura);
                 }
                 Entry::Vacant(vacant_entry) => {
-                    let mut epc = EsemplariPerCattura::new(&cattura.specie);
+                    let mut epc = EsemplariPerCattura::new(cattura.specie());
                     epc.fill_passaggio(cattura.passaggio_cattura);
                     vacant_entry.insert(epc);
                 }
@@ -473,13 +473,13 @@ fn calculate_sommatoria_x2_b_absolute(
 }
 
 fn _update_classi_eta(cl: &mut ClassiEtaSpecieNISECI, record: &RecordNISECI) {
-    if record.lunghezza < record.specie.cl_soglia_1() {
+    if record.lunghezza < record.cl_soglia_1() {
         cl.cl1 += 1;
-    } else if record.lunghezza < record.specie.cl_soglia_2() {
+    } else if record.lunghezza < record.cl_soglia_2() {
         cl.cl2 += 1;
-    } else if record.lunghezza < record.specie.cl_soglia_3() {
+    } else if record.lunghezza < record.cl_soglia_3() {
         cl.cl3 += 1;
-    } else if record.lunghezza < record.specie.cl_soglia_4() {
+    } else if record.lunghezza < record.cl_soglia_4() {
         cl.cl4 += 1;
     } else {
         cl.cl5 += 1;