+++ src/tests/model/niseci/mod.rs
@@ -40,26 +40,20 @@ fn test_calcolo_pesci_per_passaggio() {
         2.0,
     );
 
-    let record_1 = RecordNISECI {
-        specie: specie_1.clone(),
-        passaggio_cattura: 1,
-        lunghezza: 100, // in millimetri
-        peso: 100.0,    // in grammi
-    };
+    let record_1 = RecordNISECI::new(
+        &specie_1, 1, 100,   // in millimetri
+        100.0, // in grammi
+    );
 
-    let record_2 = RecordNISECI {
-        specie: specie_1.clone(),
-        passaggio_cattura: 2,
-        lunghezza: 100, // in millimetri
-        peso: 100.0,    // in grammi
-    };
+    let record_2 = RecordNISECI::new(
+        &specie_1, 2, 100,   // in millimetri
+        100.0, // in grammi
+    );
 
-    let record_3 = RecordNISECI {
-        specie: specie_1.clone(),
-        passaggio_cattura: 3,
-        lunghezza: 100, // in millimetri
-        peso: 100.0,    // in grammi
-    };
+    let record_3 = RecordNISECI::new(
+        &specie_1, 3, 100,   // in millimetri
+        100.0, // in grammi
+    );
 
     let mut c1 = vec![record_1; 20];
     let mut c2 = vec![record_2; 15];